// src/app/api/swagger/models/api-tenant-setting-dto.ts
export interface ApiTenantSettingDto {
  AllowSendToMovingPro?: boolean;
  AllowUserManagement?: boolean;
  AvailableRegions?: Array<string> | null;
  BannerHtml?: string | null;
  BusinessOperatorId?: ApiBusinessOperatorIdEnum;
  CanLinkNotTookPlaceTrade?: boolean;
  CanLinkNotTookPlaceTradeOnce?: boolean;
  CommercialOfferFromOpenPartIsEnabled?: boolean;
  CommercialOfferOwnerGuid?: string | null;
  CustomCss?: string | null;
  DenyTradePriceRaise?: boolean;
  EnableApplicationsDenyPriceRaise?: boolean;
  EnableMonetizationMarket?: boolean;
  EnableRaiseApplicationPriceMoreThanTradePrice?: boolean;
  FooterHtml?: string | null;
  HeaderHtml?: string | null;
  HeaderTitle?: string | null;
  HideApplicationDataForCustomerMode?: ApiHideApplicationDataForCustomerModeEnum;
  HideApplicationDataForParticipantMode?: ApiHideApplicationDataForParticipantModeEnum;
  HideApplicationDataOpenPartMode?: ApiHideApplicationDataModeOpenPartEnum;
  HidedSection?: Array<ApiSectionEnum> | null;
  Id?: number;
  ImmediateTradesDisable?: boolean;
  ImportExcelTemplateType?: ApiPositionsImportExcelTemplateTypeEnum;
  InstructionFileGuid?: string | null;
  IsCartEnable?: boolean;
  IsChangeSingleSupplierTradeToPriceRequest?: boolean;
  IsClarificationRequestsEnabled?: boolean;
  IsDealAgreementEnabled?: boolean;
  IsDealsOutOfEtpEnabled?: boolean;
  IsGlobalMarket?: boolean;
  IsImportOffersFromYMLEnabled?: boolean;
  IsImportPositionsFromExcelEnabled?: boolean;
  IsLandingEnable?: boolean;
  IsLandingSwitchEnabled?: boolean;
  IsManufacturerEnabled?: boolean;
  IsMyPriceListsEnabled?: boolean;
  IsNeedTradeApprove?: boolean;
  IsOfferContentComplaintsEnabled?: boolean;
  IsOfferModerationEnable?: boolean;
  IsPassportAuthEnable?: boolean;
  IsPriceCorridorEnabled?: boolean;
  IsProposalDocsRequired?: boolean;
  IsSalesTechAuthEnable?: boolean;
  IsStarOkServiceEnabled?: boolean;
  IsStoriesEnabled?: boolean;
  IsTableEditingModeEnabled?: boolean;
  IsTelegramEnabled?: boolean;
  IsZeroSingleSupplierTradeEnabled?: boolean;
  LandingBannerForCustomerLargeText?: string | null;
  LandingBannerForCustomerSmallText?: string | null;
  LandingBannerForSupplierLargeText?: string | null;
  LandingBannerForSupplierSmallText?: string | null;
  LandingMetaDescriptionContent?: string | null;
  LandingSchemeColorName?: string | null;
  LandingUrl?: string | null;
  LandingYandexMetrikaCode?: number | null;
  LogoSrc?: string | null;
  MainMenuCustomItems?: Array<ApiMenuCustomItemTemplateDto> | null;
  MaxNmz?: number | null;
  MessageKind?: ApiMessageKindAltEnum;
  MonetizationMarketWithoutNmz?: boolean;
  NewDealSignFlowForBft?: Array<number> | null;
  OrdersSideFilterStructure?: ApiSideFilterStructureDto | null;
  PriceListsSideFilterStructure?: ApiSideFilterStructureDto | null;
  RegionalProductSigns?: Array<ApiTenantRegionalProductSignSettingDto> | null;
  RequiredTradeDocumentTypes?: Array<ApiTradeDocumentTypeEnum> | null;
  ShowBankingDetailsOnCreateApplicationForm?: boolean;
  ShowTradeCustomerInfo?: boolean;
  StatisticsAndAnalyticsReportUrl?: string | null;
  TenantFlowType?: ApiTenantFlowTypeAltEnum;
  TenantHosts?: Array<string> | null;
  ZmoPersonalAccountUrl?: string | null;
}